requirements.txt
streamlit
cloudscraper
pandas
trafilatura
google-generativeai
lxml
gspread
google-auth
python-docx
openpyxl